pubsub-to-image/requirements.txt
google-cloud-storage
Pillow
